+++ src/validators/swagger_request_models.py
@@ -1080,11 +1080,6 @@ class JobPostModel(BaseModel):
         'automatically unless an ambiguous occurs',
     )
 
-    application_id: str = Field(
-        None,
-        description='Application ID with credentials for this job',
-    )
-
     dojo_product: str = Field(
         None,
         description='Defect Dojo product name to which the results will be '
@@ -1633,6 +1628,7 @@ class ProjectGetReportModel(BaseModel):
         ]
     ] = Field(default_factory=set)
     receivers: set[str] = Field(default_factory=set)
+    include_linked: bool = False
     # attempt: SkipJsonSchema[int] = 0
     # execution_job_id: SkipJsonSchema[str] = Field(None)